Move 'curl' from build dependencies to permanent dependencies
In this commit, we've moved 'curl' from the '.build-deps' to the base apk installation command. This change has been made as our Docker container requires 'curl' during the runtime, not just during the build phase.

Please note, this change might increase the size of the Docker image, as 'curl' is now a permanent dependency and won't be cleaned up once the build phase ends.